refactor: code organization wip
