+++ dev-client/src/components/home/BackdropComponent.tsx
@@ -1,0 +1,6 @@
+import {BottomSheetBackdrop} from '@gorhom/bottom-sheet';
+import {BottomSheetBackdropProps} from '@gorhom/bottom-sheet';
+
+export const BackdropComponent = (props: BottomSheetBackdropProps) => (
+  <BottomSheetBackdrop {...props} appearsOnIndex={0} disappearsOnIndex={-1} />
+);
