+++ dev-client/src/components/home/BottomSheet.tsx
@@ -16,57 +16,18 @@
  */
 
 import BottomSheet, {BottomSheetFlatList} from '@gorhom/bottom-sheet';
-import {
-  Box,
-  Row,
-  Heading,
-  Text,
-  useTheme,
-  Column,
-  FormControl,
-  Select,
-  Spinner,
-} from 'native-base';
-import {
-  Dispatch,
-  SetStateAction,
-  forwardRef,
-  useCallback,
-  useMemo,
-  useState,
-} from 'react';
-import {Linking} from 'react-native';
+import {Box, Row, Heading, Text, useTheme, Column, Spinner} from 'native-base';
+import {forwardRef, useCallback, useMemo} from 'react';
 import {useSelector} from 'terraso-mobile-client/model/store';
 import {Site} from 'terraso-client-shared/site/siteSlice';
-import {Trans, useTranslation} from 'react-i18next';
-import {LinkNewWindowIcon} from 'terraso-mobile-client/components/common/Icons';
+import {useTranslation} from 'react-i18next';
 import {SiteCard} from 'terraso-mobile-client/components/sites/SiteCard';
 import {BottomSheetMethods} from '@gorhom/bottom-sheet/lib/typescript/types';
-import {SearchBar} from 'terraso-mobile-client/components/common/search/SearchBar';
-import {SiteFilter} from 'terraso-mobile-client/components/sites/filter.hooks';
-import {ProjectSelect} from 'terraso-mobile-client/components/projects/ProjectSelect';
-import {USER_ROLES} from 'terraso-mobile-client/constants';
-
-const EmptySiteMessage = () => {
-  const {t} = useTranslation();
-
-  return (
-    <Text px="17px" variant="body1">
-      <Trans
-        i18nKey="site.empty.info"
-        components={{
-          icon: <LinkNewWindowIcon />,
-        }}>
-        <Text
-          underline
-          onPress={() => Linking.openURL(t('site.empty.link_url'))}
-          color="primary.main">
-          link_text
-        </Text>
-      </Trans>
-    </Text>
-  );
-};
+import {
+  SiteSearchBarProps,
+  SiteSearchBar,
+} from 'terraso-mobile-client/components/home/SiteSearchBarProps';
+import {EmptySiteMessage} from 'terraso-mobile-client/components/home/EmptySiteMessage';
 
 type Props = {
   sites: Site[];
@@ -137,70 +98,3 @@ export const SiteListBottomSheet = forwardRef<BottomSheetMethods, Props>(
     );
   },
 );
-
-type SiteSearchBarProps = {
-  query: string;
-  setQuery: (query: string) => void;
-  filter: SiteFilter;
-  setFilter: (filter: SiteFilter) => void;
-};
-const SiteSearchBar = ({
-  query,
-  setQuery,
-  filter,
-  setFilter,
-}: SiteSearchBarProps) => {
-  const {t} = useTranslation();
-  const numFilters = Object.values(filter).filter(
-    value => value !== undefined,
-  ).length;
-  const [tempFilter, setTempFilter] = useState(filter);
-
-  return (
-    <SearchBar
-      query={query}
-      setQuery={setQuery}
-      numFilters={numFilters}
-      placeholder={t('site.search.placeholder')}
-      FilterOptions={
-        <SiteFilterModal filter={tempFilter} setFilter={setTempFilter} />
-      }
-      onApplyFilter={() => setFilter(tempFilter)}
-    />
-  );
-};
-
-type SiteFilterModalProps = {
-  filter: SiteFilter;
-  setFilter: Dispatch<SetStateAction<SiteFilter>>;
-};
-const SiteFilterModal = ({filter, setFilter}: SiteFilterModalProps) => {
-  const {t} = useTranslation();
-  return (
-    <Column>
-      <FormControl>
-        <FormControl.Label>
-          {t('site.search.filter_projects')}
-        </FormControl.Label>
-        <ProjectSelect
-          projectId={filter.projectId}
-          setProjectId={projectId => setFilter(prev => ({...prev, projectId}))}
-        />
-      </FormControl>
-      <FormControl>
-        <FormControl.Label>{t('site.search.filter_role')}</FormControl.Label>
-        <Select
-          selectedValue={filter.role}
-          onValueChange={role => setFilter(prev => ({...prev, role}) as any)}>
-          {USER_ROLES.map(role => (
-            <Select.Item
-              label={t(`site.role.${role}`)}
-              value={role}
-              key={role}
-            />
-          ))}
-        </Select>
-      </FormControl>
-    </Column>
-  );
-};
