+++ dev-client/src/components/projects/ProjectPreviewCard.tsx
@@ -18,7 +18,7 @@
 import {Badge, HStack, Heading, Text} from 'native-base';
 import {useTranslation} from 'react-i18next';
 import {useCallback} from 'react';
-import {useNavigation} from 'terraso-mobile-client/screens/AppScaffold';
+import {useNavigation} from 'terraso-mobile-client/screens/useNavigation';
 import {Card} from 'terraso-mobile-client/components/common/Card';
 import {Icon} from 'terraso-mobile-client/components/common/Icons';
 import {Project} from 'terraso-client-shared/project/projectSlice';
